Refactoring and partial made player for dictations

+++ nps/src/main/java/com/dpcat237/nps/adapter/FeedsAdapter.java
@@ -75,7 +75,7 @@ public class FeedsAdapter extends BaseAdapter {
         Feed feed =getItem(position);
 
         holder.title.setText(feed.getTitle());
-        Long unread = feed.getUnreadCount();
+        Integer unread = feed.getUnreadCount();
         if (unread > 0) {
             holder.count.setText(unread.toString());
         } else {
